Refactor cuckoo
