+++ consensus/crypto/cuckaroo/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cuckoo"
+name = "cuckaroo"
 version = "0.1.0"
 authors = ["Starcoin Core Dev <dev@starcoin.org>"]
 license = "Apache-2.0"
@@ -10,6 +10,13 @@ edition = "2018"
 logger = { package = "libra-logger", path = "../../../common/logger" }
 byteorder = "1.3.1"
 blake2-rfc = "0.2"
+num = "0.2"
+num-bigint = "0.2"
+failure = "0.1"
+failure_derive = "0.1"
+rand = "0.6"
+serde = "1"
+serde_derive = "1"
 
 [build-dependencies]
 cc = "1.0"
