+++ consensus/crypto/cuckaroo/src/error.rs
