+++ consensus/crypto/cuckaroo/src/lib.rs
@@ -1,0 +1,10 @@
+pub use cuckaroo::new_cuckaroo_ctx;
+pub use types::{PoWContext, Proof as CuckarooProof, PROOF_SIZE};
+
+#[macro_use]
+mod common;
+mod cuckaroo;
+mod error;
+mod siphash;
+
+mod types;
