+++ consensus/crypto/cuckaroo/src/siphash.rs
