--- consensus/crypto/cuckoo/build.rs
@@ -1,9 +1,0 @@
-fn main() {
-    cc::Build::new()
-        .file("src/cuckoo.c")
-        .include("src/")
-        .flag("-O3")
-        // .flag("-march=native")
-        .static_flag(true)
-        .compile("libmean.a");
-}
